% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{as_params_pkg}
\alias{as_params_pkg}
\title{Generate a character vector of parameters for package options}
\usage{
as_params_pkg(envir = NULL, include = NULL, exclude = NULL)
}
\arguments{
\item{envir}{The environment containing the options. Can be NULL (default,
uses the calling environment), a string (package name), or an environment object.}

\item{include}{A character vector of option names to include. If NULL (default),
all options are included.}

\item{exclude}{A character vector of option names to exclude. Default is NULL.}
}
\value{
A character vector of parameter descriptions, where each element is a string
describing an option.
}
\description{
This function generates a character vector of parameters for options defined using
the custom options functions. The resulting vector can be used in documentation
or for other purposes.
}
\examples{
\dontrun{
# Assuming some options are defined in the current environment
params <- as_params_pkg()
cat(params, sep = "\n")

# For a specific package
params <- as_params_pkg("mypackage")
cat(params, sep = "\n")

# Include only specific options
params <- as_params_pkg(include = c("option1", "option2"))
cat(params, sep = "\n")

# Exclude specific options
params <- as_params_pkg(exclude = c("option3", "option4"))
cat(params, sep = "\n")
}

}
