% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{as_roxygen_docs_pkg}
\alias{as_roxygen_docs_pkg}
\title{Generate Roxygen documentation for package options}
\usage{
as_roxygen_docs_pkg()
}
\value{
A character vector containing Roxygen documentation for the
\code{set_pkg_options()} function, including descriptions of all available options.
}
\description{
This function generates Roxygen documentation for all package options
defined in the global environment. It creates documentation for a
\code{set_pkg_options()} function, which allows users to set these options.
}
\examples{
# Assuming you have defined some options in your package
cat(as_roxygen_docs_pkg(), sep = "\n")

}
