% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmos.R
\name{create_atmos_package}
\alias{create_atmos_package}
\title{Create new repository based on template}
\usage{
create_atmos_package(
  repo,
  template = "NN-OpenSource/Rtemplate",
  use_standards = TRUE
)
}
\arguments{
\item{repo}{New GitHub repository (owner/name)}

\item{template}{Template GitHub repository (owner/name)}

\item{use_standards}{\link{logical} apply settings and rules from the template repository?}
}
\value{
\link{invisible} output from the GitHUb Api when creating the repository
}
\description{
Create new repository based on template
}
