% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_verbosity_level.R
\name{get_envname_options}
\alias{get_envname_options}
\title{Get the value of the \code{opt_source} for an environment variable based on \code{options} package version}
\usage{
get_envname_options()
}
\value{
\code{character} with the value of the \code{opt_source} for an environment variable
}
\description{
See breaking change in version 0.2.0 in the changelog here:
https://dgkf.github.io/options/news/index.html. Function gives the value that is
returned by \code{\link[=opt_source]{opt_source()}}. Is used inside the \code{\link[=get_verbosity_level]{get_verbosity_level()}} function.
}
\examples{
get_envname_options()
}
