% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{get_option_spec_pkg}
\alias{get_option_spec_pkg}
\title{Get an option specification.}
\usage{
get_option_spec_pkg(option_name, envir = parent.frame(), print_spec = TRUE)
}
\arguments{
\item{option_name}{The name of the option.}

\item{envir}{The environment to search for the option.}

\item{print_spec}{Whether to print the option specification.}
}
\value{
An option specification object.
}
\description{
This function retrieves an option specification.
}
