% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_verbosity_level.R
\name{get_package_name_and_verbosity}
\alias{get_package_name_and_verbosity}
\title{Get Package Name and Verbosity Level from Global Options}
\usage{
get_package_name_and_verbosity()
}
\value{
A list with two elements:
\item{package_name}{A character string containing the name of the current package.}
\item{verbosity_level}{A character string indicating the verbosity level,
or NULL if not specified or if invalid. Possible values are "quiet", "minimal", "verbose", and "debug".}
}
\description{
This function retrieves the current package name and searches the global environment
for an option that specifies the verbosity level for this package.
}
\details{
The function determines the current package name, then looks for a global option
named "\link{package_name}.verbosity_level". If found, it returns the package name and verbosity level.
The verbosity level is NULL if not specified or if invalid.
}
\examples{
# Assuming this function is called from within the "zephyr" package:
# Set a global option for the current package with verbose verbosity level
options(zephyr.verbosity_level = "verbose")
get_package_name_and_verbosity()  # Returns list(package_name = "zephyr", verbosity_level = "verbose")

}
