% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_option.R
\name{get_package_option}
\alias{get_package_option}
\title{Get Package Option}
\usage{
get_package_option(option_name, pkg_env)
}
\arguments{
\item{option_name}{A character string specifying the name of the option to search for.}

\item{pkg_env}{The package environment to search in. This should be an environment object.}
}
\value{
The name of the matching option if found, otherwise NULL.
}
\description{
This function searches for a specified option within a package environment.
It looks for objects that either end with the given option name or match it exactly.
The search is case-insensitive.
}
\details{
The function first searches in the main package environment. If the option is not found there,
it checks in the \code{.options} environment if it exists within the package environment.
If multiple matching options are found, it returns the first one and issues a warning.
}
\note{
This function does not return the value of the option, only its name.
To get the value, use \code{opt_pkg(result, envir = pkg_env)} on the returned result.
}
\examples{
\dontrun{
# Assuming 'my_package' is loaded and has an option 'verbose'
pkg_env <- asNamespace("my_package")
option_name <- get_package_option("verbosity_level", pkg_env)
if (!is.null(option_name)) {
  option_value <- opt_pkg(option_name, envir = pkg_env)
  print(paste("Value of", option_name, "is", option_value))
}
}

}
