% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opt.R
\name{ns_of_call}
\alias{ns_of_call}
\title{Get the namespace of function in the call stack}
\usage{
ns_of_call(which = 1)
}
\arguments{
\item{which}{Passed onto \code{\link[base:sys.parent]{base::sys.call()}}}
}
\value{
The namespace of the package that function in \code{sys.call(which)}
belongs to
}
\description{
Get the namespace of the function somewhere in the call stack.
Default behavior will get the namespace of the package from which the top
function in the stack comes from
}
\examples{
# Get the namespace of package from which the top function in the call stack
# comes from
ns_of_call()
}
