% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{opt_pkg}
\alias{opt_pkg}
\title{Get the value of an option.}
\usage{
opt_pkg(option_name, default = NULL, envir = parent.frame())
}
\arguments{
\item{option_name}{The name of the option.}

\item{default}{The default value to return if the option is not found.}

\item{envir}{The environment to search for the option.}
}
\value{
The value of the option.
}
\description{
This function retrieves the value of an option.
}
