% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{opt_source_pkg}
\alias{opt_source_pkg}
\title{Determine the source of an option value}
\usage{
opt_source_pkg(x, envir = parent.frame())
}
\arguments{
\item{x}{The name of the option or an option specification object.}

\item{envir}{The environment in which to look for the option.}
}
\value{
A character string indicating the source of the option value.
}
\description{
This function determines the source of an option value (option, environment variable, or default).
}
\examples{
define_option_pkg("my_option", default = 42)
opt_source_pkg("my_option")

}
