% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{option_spec_pkg}
\alias{option_spec_pkg}
\title{Create an option specification.}
\usage{
option_spec_pkg(
  name,
  default = NULL,
  desc = NULL,
  option_name = NULL,
  envvar_name = NULL,
  option_fn = NULL,
  envvar_fn = NULL,
  quoted = FALSE,
  eager = FALSE,
  envir = parent.frame(),
  print_spec = TRUE
)
}
\arguments{
\item{name}{The name of the option.}

\item{default}{The default value of the option.}

\item{desc}{A description of the option.}

\item{option_name}{The name of the option in the options list.}

\item{envvar_name}{The name of the environment variable.}

\item{option_fn}{A function to apply to the option value.}

\item{envvar_fn}{A function to apply to the environment variable value.}

\item{quoted}{Whether the default value is quoted.}

\item{eager}{Whether to evaluate the default value eagerly.}

\item{envir}{The environment to evaluate the default value in.}

\item{print_spec}{Whether to print the option specification.}
}
\value{
An option specification object (invisibly).
}
\description{
This function creates an option specification.
}
