% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{remove_option_pkg}
\alias{remove_option_pkg}
\title{Remove a package-specific option}
\usage{
remove_option_pkg(option, envir = parent.frame())
}
\arguments{
\item{option}{A character string specifying the name of the option to remove.}

\item{envir}{The environment where the \code{.options} environment is located.
Default is \code{parent.frame()}.}
}
\value{
Invisible NULL. The function is called for its side effect of removing the option.
}
\description{
This function removes a package-specific option that was previously set using
the \code{define_option_pkg} function.
}
\details{
The function removes the specified option from the \code{.options} environment
within the given environment. If the \code{.options} environment or the specified
option doesn't exist, the function will do nothing.
}
\examples{
# Assuming we have previously set an option:
# define_option_pkg("my_option", default = "value")

# To remove the option:
remove_option_pkg("my_option")

}
