% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_option}
\alias{set_option}
\title{Set value of package option}
\usage{
set_option(name, value, .envir = parent.frame())
}
\arguments{
\item{name}{\verb{[character(1)]} Name of the option}

\item{value}{Value to set for the option. Use \code{NULL} to unset previously set R option.}

\item{.envir}{Environment in which the option is defined.
Default is suitable for use inside your package.}
}
\value{
Invisible previous value of the option (from \code{\link[=get_option]{get_option()}})
}
\description{
Sets the value of a zephyr option. This is a convenience wrapper around
\code{\link[=options]{options()}} that handles the option naming convention.

The option is set as \verb{\{pkgname\}.\{name\}} (lowercase).
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
set_option(name = "my_option", value = "my_value", .envir = "my_pkg")
\dontshow{\}) # examplesIf}
}
