% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity_level.R
\name{set_verbosity_level}
\alias{set_verbosity_level}
\title{Set verbosity level}
\usage{
set_verbosity_level(
  level = c("quiet", "minimal", "verbose", "debug"),
  .envir = "zephyr"
)
}
\arguments{
\item{level}{\verb{[character(1)]} Verbosity level to set. Must be one of
\code{"quiet"}, \code{"minimal"}, \code{"verbose"}, or \code{"debug"}.}

\item{.envir}{Environment in which the options are defined.
Default \code{zephyr} is suitable to affect all packages using zephyr.}
}
\value{
Invisible \verb{[character(1)]} of the previous verbosity level.
}
\description{
Sets the verbosity level for a package. This is a convenience wrapper around
\code{\link[=set_option]{set_option()}} for the \code{verbosity_level} option.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Set verbosity level to minimal for zephyr
set_verbosity_level("minimal")
\dontshow{\}) # examplesIf}
}
