% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{spinner}
\alias{spinner}
\title{Display a CLI spinner while running a function}
\usage{
spinner(x = NULL, msg = "Processing... ")
}
\arguments{
\item{x}{A function to execute while the spinner is running.}

\item{msg}{Message to display alongside the spinner.#'}
}
\description{
Runs a function while displaying a CLI spinner animation.
The spinner appears while the given function is executing, and
disappears when it finishes (whether successfully or not).
}
\examples{
\dontrun{
# Execute a function with a spinner
spinner(function() Sys.sleep(1), msg = "Waiting with spinner...")

# Any return value is returned from the function you provided:
result <- spinner(function() {
  Sys.sleep(0.5)
  1 + 1
}, msg = "Calculating 1+1...")
print(result)
}
}
