% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{start_spinner}
\alias{start_spinner}
\title{Start a CLI spinner}
\usage{
start_spinner(msg = "Processing... ")
}
\arguments{
\item{msg}{Message to display alongside spinner}
}
\value{
A list containing spinner context information
}
\description{
Start a spinner process to indicate work in progress on the CLI.
}
\examples{
\dontrun{
# Start a spinner, do some work, then stop it
ctx <- start_spinner("Starting some long task...")
Sys.sleep(1)
stop_spinner(ctx)

# It is recommended to always pair with stop_spinner()
# to properly clean up resources.
}
}
