% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmos.R
\name{use_atmos_pages}
\alias{use_atmos_pages}
\title{Enable GitHUb Pages and get the url}
\usage{
use_atmos_pages(repo)
}
\arguments{
\item{repo}{GitHub repository (owner/name)}
}
\value{
\link{invisible} output from the GitHUb Api when enabling Pages
}
\description{
Enable GitHUb Pages and get the url
}
