% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmos.R
\name{use_atmos_rulesets}
\alias{use_atmos_rulesets}
\title{Apply rule sets from template repository}
\usage{
use_atmos_rulesets(repo, template = "NN-OpenSource/Rtemplate")
}
\arguments{
\item{repo}{New GitHub repository (owner/name)}

\item{template}{Template GitHub repository (owner/name)}
}
\value{
\link{invisible} output from the GitHUb Api when creating the repository
}
\description{
Apply rule sets from template repository
}
