% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmos.R
\name{use_atmos_settings}
\alias{use_atmos_settings}
\title{Apply settings from template repository and updates link to homepage}
\usage{
use_atmos_settings(repo, homepage, template = "NN-OpenSource/Rtemplate")
}
\arguments{
\item{repo}{New GitHub repository (owner/name)}

\item{homepage}{URL of package webpage}

\item{template}{Template GitHub repository (owner/name)}
}
\value{
\link{invisible} output from the GitHUb Api when creating the repository
}
\description{
Apply settings from template repository and updates link to homepage
}
