% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity_level.R
\name{use_verbosity_level}
\alias{use_verbosity_level}
\title{Use verbosity level}
\usage{
use_verbosity_level()
}
\description{
Sets up the \code{verbosity_level} option in your package and adds the
documentation.
}
