% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_zephyr.R
\name{use_zephyr}
\alias{use_zephyr}
\title{Use zephyr options and verbosity levels}
\usage{
use_zephyr()
}
\description{
Utility function to set up the following in your package:
}
\details{
\enumerate{
\item Use of \code{verbosity_level}
\item Use of package specific options
\item Documents all options
}
}
