% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity_level.R
\name{validate_verbosity_level}
\alias{validate_verbosity_level}
\title{Validate verbosity level}
\usage{
validate_verbosity_level(level)
}
\arguments{
\item{level}{The verbosity level to validate}
}
\value{
A valid verbosity level
}
\description{
Validate verbosity level
}
\keyword{internal}
