% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{with_spinner}
\alias{with_spinner}
\title{\code{spinner} wrapper to avoid LHS priority eval limitations with \verb{|>}}
\usage{
with_spinner(expr, msg = "Running: {.expr}")
}
\arguments{
\item{expr}{The expression to evaluate}

\item{msg}{The message to display alongside the spinner default is \code{'Running: {.expr}'}}
}
\value{
The result of evaluating \code{expr}
}
\description{
This function evaluates an expression while displaying a spinner animation
with a custom message.
}
\examples{
#' # Simple delay with spinner
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
with_spinner(Sys.sleep(2), "Waiting for 2 seconds")

# Complex expressions
with_spinner({
  Sys.sleep(1)
  Sys.sleep(1)
  "Result"
}, "Processing complex operation")
\dontshow{\}) # examplesIf}
}
