% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msg.R
\name{write_msg}
\alias{write_msg}
\title{Write messages based on verbosity level}
\usage{
write_msg(
  message,
  verbosity_levels = c("verbose", "debug"),
  msg_fun = cli::cli_h1,
  env = rlang::ns_env(),
  ...
)
}
\arguments{
\item{message}{\code{character} of message to write}

\item{verbosity_levels}{\code{character} vector of levels of verbosity for which}

\item{msg_fun}{\code{function} taking \code{message} as first argument. Usually a
\code{cli_...} function}

\item{...}{Additional arguments passed to \code{msg_fun}}

\item{x}{Code to be printed based on verbosity level}
}
\description{
Write based on verbosity level set as option in package
(see \link{options}),
}
\examples{
# Use the `write_msg` function to fx. create headlines for big steps in the
# TFL creation process
write_msg("Initialising {.field ingest_tfl} object",
          verbosity_levels = c("verbose", "debug"),
          msg_fun = cli::cli_h1)

ingest_init(pharmaverseadam::adae, filter = SAFFL == "Y")

# Use the `with_verbose` function to
data \%>\%
  dplyr::filter(SAFFL == "Y") \%>\%
  with_verbose()

}
