% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{define_option_pkg}
\alias{define_option_pkg}
\title{Define an option}
\usage{
define_option_pkg(
  option,
  default = NULL,
  desc = NULL,
  option_name = NULL,
  envvar_name = NULL,
  option_fn = NULL,
  envvar_fn = NULL,
  quoted = FALSE,
  eager = FALSE,
  envir = parent.frame(),
  print_spec = TRUE
)
}
\arguments{
\item{option}{The name of the option to define.}

\item{default}{The default value of the option.}

\item{desc}{A description of the option.}

\item{option_name}{The name of the R option (defaults to \code{name}).}

\item{envvar_name}{The name of the environment variable (defaults to \verb{R_[uppercase name]}).}

\item{option_fn}{A function to process the option value.}

\item{envvar_fn}{A function to process the environment variable value.}

\item{quoted}{Whether the option value should be quoted.}

\item{eager}{Whether to evaluate the default value eagerly.}

\item{envir}{The environment in which to define the option. Defaults to the parent frame.}

\item{print_spec}{Whether to print the option specification.}
}
\value{
An invisible option specification object.
}
\description{
This function defines an option for a package. It attempts to set the option
in the package environment. If that fails (e.g., due to a locked namespace),
it falls back to setting a global option with a "zephyr." prefix.
}
\examples{
define_option_pkg("verbosity_level", default = "verbose", desc = "Verbosity level for output")

}
