% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{envvar_str_split_pkg}
\alias{envvar_str_split_pkg}
\title{Split an environment variable or option string into a vector}
\usage{
envvar_str_split_pkg(delim = ",", envir = NULL)
}
\arguments{
\item{delim}{The delimiter to use for splitting the string. Default is ",".}

\item{envir}{The environment in which to look for the option. Can be NULL (default, uses .GlobalEnv),
a string (package name), or an environment.}
}
\value{
A function that takes parameters:
\itemize{
\item x: The name of the option or environment variable to check
\item ...: Additional arguments (currently unused)
}
}
\description{
This function returns another function that splits a given string (from an environment variable or option)
into a vector using a specified delimiter, and trims whitespace.
}
\examples{
split_fn <- envvar_str_split_pkg()
Sys.setenv(R_MY_LIST_OPTION = "item1,item2,item3")
split_fn("my_list_option")  # Returns c("item1", "item2", "item3")

Sys.setenv(MY_ENV_LIST = "x|y|z")
split_fn <- envvar_str_split_pkg(delim = "|")
split_fn("MY_ENV_LIST")  # Returns c("x", "y", "z")

}
