% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_option_spec_pkg}
\alias{get_option_spec_pkg}
\title{Get the option specification}
\usage{
get_option_spec_pkg(x, envir = NULL, print_spec = FALSE)
}
\arguments{
\item{x}{The name of the option.}

\item{envir}{The environment in which to look for the option. Can be NULL (default), a string (package name), or an environment.}

\item{print_spec}{Whether to print the option specification.}
}
\value{
The option specification object, or NULL if not found.
}
\description{
This function retrieves the option specification for a given option name.
}
\examples{
define_option_pkg("my_option", default = 42)
get_option_spec_pkg("my_option")
get_option_spec_pkg("my_option", envir = "zephyr")

}
