% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity_level.R
\name{get_package_name_and_env_verbosity}
\alias{get_package_name_and_env_verbosity}
\title{Get Package Name and Environment Variable Verbosity Level}
\usage{
get_package_name_and_env_verbosity()
}
\value{
A list with three elements:
\item{package_name}{A character string with the detected package name, or NULL if in global environment}
\item{env_var_name}{A character string with the name of the environment variable}
\item{verbosity_level}{A character string with the verbosity level, or NULL if not set or invalid}
}
\description{
This function determines the calling package name and checks for a corresponding
environment variable that sets the verbosity level.
}
\examples{
result <-  get_package_name_and_env_verbosity()
print(result)

}
