% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_verbosity_level.R
\name{get_package_name_and_verbosity}
\alias{get_package_name_and_verbosity}
\title{Get Package Name and Verbosity Level}
\usage{
get_package_name_and_verbosity()
}
\value{
A list with two elements:
\item{package_name}{A character string with the detected package name, or NULL if in global environment}
\item{verbosity_level}{A character string with the verbosity level, or NULL if not set}
}
\description{
This function determines the calling package name and checks for a corresponding
verbosity level option. It works in package environments, devtools environments,
and the global environment.
}
\examples{
result <-  get_package_name_and_verbosity()
print(result)

}
