% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_verbosity_level.R
\name{get_verbosity_level}
\alias{get_verbosity_level}
\title{Get \code{verbosity_level} option, prioritizing different sources.}
\usage{
get_verbosity_level(env = parent.frame())
}
\arguments{
\item{env}{Environment to get the option from.}
}
\value{
Value of the option.
}
\description{
This function retrieves the value of the \code{verbosity_level} option, prioritizing
different sources according to the following rules:
}
\details{
\enumerate{
\item If the option is not defined for the specified environment, use the option
set on the "zephyr" level (using the default if nothing is done).
\item If the option is set on the "zephyr" level, use that one.
\item If the option is set on the "env" level, use that one.
\item If the option is not set as an option on the package level, and the "zephyr"
option is set using an environment variable, use that one.
\item If the "zephyr" level option is not set, use the option set on the package
level, whether default or environment variable.
}
}
\examples{
# Define the option in the "zephyr" namespace

# Get the verbosity level in the current environment
get_verbosity_level()

# Get the verbosity level in a specific environment
env <- new.env()
get_verbosity_level(env)
}
