% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_verbosity_level.R
\name{get_verbosity_level}
\alias{get_verbosity_level}
\title{Get \code{verbosity_level} option using the \code{options} package, allowing a global option}
\usage{
get_verbosity_level(env = parent.frame())
}
\arguments{
\item{env}{Environment to get the option from (set with the \code{options} package)}
}
\value{
Value of the option
}
\description{
Get option using the \code{options} package, where the
function allows a global option to overwrite individual package options.
}
\examples{
# Set a package-level verbosity option
define_option_pkg("verbosity_level", default = "verbose",
                  desc = "Set the verbosity level for messages")

# Get the verbosity level (should return "verbose")
get_verbosity_level()

# Set a global option
options(zephyr.verbosity_level = "debug")

# Get the verbosity level (should now return "debug")
get_verbosity_level()

# Set an environment variable
Sys.setenv(R_ZEPHYR_VERBOSITY_LEVEL = "quiet")

# Get the verbosity level (should still return "debug" due to option priority)
get_verbosity_level()

# Remove the global option
options(zephyr.verbosity_level = NULL)

# Get the verbosity level (should now return "quiet" from the environment variable)
get_verbosity_level()
}
