% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msg.R
\name{ns_of_call}
\alias{ns_of_call}
\title{Get the namespace of function in the call stack}
\usage{
ns_of_call(which = 0)
}
\arguments{
\item{which}{Passed onto \code{\link[base:sys.parent]{base::sys.call()}}}
}
\value{
The namespace of the package that function in \code{sys.call(which)}
belongs to
}
\description{
Get the namespace of the function somewhere in the call stack.
Default behavior will get the namespace of the package of the function that
called the function from which this is called. Used inside \code{\link[=write_msg]{write_msg()}}, so
the verbosity level set in the package, which uses \code{\link[=write_msg]{write_msg()}}, will be used
}
\examples{
# Get the namespace of the function that this is in
zephyr:::ns_of_call()

\dontrun{
# Get the namespace of package of the function that called this function
ns_of_call(which = -1)

# Use case of write_msg function: Get the namespace of package of the
# function that called write_msg
write_msg <- function(...) {
 ...
 ns_of_call(which = -2)
 ...
}
}

}
