% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{opt_source_pkg}
\alias{opt_source_pkg}
\title{Determine the source of an option value.}
\usage{
opt_source_pkg(x, envir = parent.frame())
}
\arguments{
\item{x}{An option specification or the name of an option.}

\item{envir}{The environment to search for the option.}
}
\value{
The source of the option value ("option", "envvar", "default", or NA).
}
\description{
This function determines the source of an option value.
}
