% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-options.R
\name{opt_source_pkg}
\alias{opt_source_pkg}
\title{Determine the source of an option value}
\usage{
opt_source_pkg(spec, envir = NULL)
}
\arguments{
\item{spec}{An option specification object.}

\item{envir}{The environment in which to look for the option. Can be NULL (default), a string (package name), or an environment.}
}
\value{
A character string indicating the source of the option value: "package", "option", "envvar", or "default".
}
\description{
This function determines the source of an option value (package environment, global option, environment variable, or default).
}
\examples{
spec <- get_option_spec_pkg("verbosity_level", envir = "zephyr")
opt_source_pkg(spec, envir = "zephyr")

}
