% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{report_checkmate_assertations}
\alias{report_checkmate_assertations}
\title{Report collection of assertations}
\usage{
report_checkmate_assertations(collection, msg = "Invalid input(s):")
}
\arguments{
\item{collection}{A collection of assertations created with \code{\link[checkmate:AssertCollection]{checkmate::makeAssertCollection()}}}

\item{msg}{\code{\link[=character]{character()}} Header of the error message if any assertations failed}
}
\description{
Improved reporting of \code{AssertColellections} created with the \code{\link[checkmate:AssertCollection]{checkmate::makeAssertCollection()}} package
using \code{\link[cli:cli_abort]{cli::cli_abort()}} instead of \code{\link[checkmate:AssertCollection]{checkmate::reportAssertions()}} in order to provide a more informative error message.

THe function is inteded to be used inside a function that performs assertations on its input arguments. See below for an example.
}
\examples{
add_numbers <- function(a, b) {
  collection <- checkmate::makeAssertCollection()
  checkmate::assert_numeric(x = a, add = collection)
  checkmate::assert_numeric(x = b, add = collection)
  report_checkmate_assertations(collection)
  return(a + b)
}
\donttest{
add_numbers(1, 2)
add_numbers(1, "b")
add_numbers("a", "b")
}
}
