% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_zephyr.R
\name{use_zephyr}
\alias{use_zephyr}
\title{Use zephyr options and verbosity levels}
\usage{
use_zephyr()
}
\value{
\code{invisible(TRUE)}
}
\description{
Utility function to set up the use of zephyr options and
\link{verbosity_level} in your package.

Creates the file \code{R/{pkgname}-options.R} with boiler plate code to setup
and document options.

This code also creates an package specific \code{verbosity_level} option,
enabling you to control the verbosity of your package functions using
the \link{msg} functions.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
use_zephyr()
\dontshow{\}) # examplesIf}
}
