% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msg.R
\name{write_msg}
\alias{write_msg}
\title{Write messages based on verbosity level}
\usage{
write_msg(
  message,
  levels_to_write = c("verbose", "debug"),
  msg_fun = cli::cli_h1,
  opt_name = "verbosity_level",
  global_opt_name = paste0("atmos.", opt_name),
  ...
)
}
\arguments{
\item{message}{\code{character} of message to write}

\item{levels_to_write}{\code{character} vector of levels of verbosity for which
to display the message}

\item{msg_fun}{\code{function} taking \code{message} as first argument. Usually a
\code{cli_...} function}

\item{opt_name}{\code{character} name of the option. Passed to \code{\link[=get_opt]{get_opt()}}}

\item{global_opt_name}{\code{character} name of the global option which, if set,
will overwrite the package level option. Passed to \code{\link[=get_opt]{get_opt()}}}

\item{...}{Additional arguments passed to \code{msg_fun}}
}
\description{
Write based on verbosity level set as option in package
}
\examples{
# Use the `write_msg` function to fx. create headlines for big steps in the
# TFL creation process
write_msg("Initialising {.field ingest_tfl} object",
          levels_to_write = c("verbose", "debug"),
          msg_fun = cli::cli_h1)

ingest_init(pharmaverseadam::adae, filter = SAFFL == "Y")

# Use the `with_verbose` function to
data \%>\%
  dplyr::filter(SAFFL == "Y") \%>\%
  with_verbose()

}
