% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-package.R
\docType{package}
\name{zephyr-package}
\alias{zephyr}
\alias{zephyr-package}
\title{zephyr: Structured Messages and Options}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a structured framework for consistent user communication and configuration management for package developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://novonordisk-opensource.github.io/zephyr/}
  \item \url{https://github.com/novonordisk-opensource/zephyr}
  \item Report bugs at \url{https://github.com/novonordisk-opensource/zephyr/issues}
}

}
\author{
\strong{Maintainer}: Aksel Thomsen \email{oath@novonordisk.com}

Authors:
\itemize{
  \item Mathias Lerbech Jeppesen \email{nmlj@novonordisk.com}
  \item Cervan Girard \email{cgid@novonordisk.com}
  \item Kristian Troejelsgaard \email{ktqn@novonordisk.com}
  \item Lovemore Gakava \email{lvgk@novonordisk.com}
  \item Steffen Falgreen Larsen \email{sffl@novonordisk.com}
  \item Vladimir Obucina \email{vlob@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
