% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-package.R
\docType{package}
\name{zephyr-package}
\alias{zephyr}
\alias{zephyr-package}
\title{zephyr: Template repository for new ATMOS R packages}
\description{
Defines common standards, workflows, and actions for new R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jubilant-adventure-yr95p3g.pages.github.io/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
