% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zephyr-package.R
\docType{package}
\name{zephyr-package}
\alias{zephyr}
\alias{zephyr-package}
\title{zephyr: Tools for Informing Users Based on Package Options}
\description{
Tools for helping developers of R packages inform their users. Package enables writing messages based on package level options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jubilant-adventure-yr95p3g.pages.github.io/}
}

}
\author{
\strong{Maintainer}: Aksel Thomsen \email{oath@novonordisk.com}

Authors:
\itemize{
  \item Mathias Lerbech Jeppesen \email{nmlj@novonordisk.com}
  \item Cervan Girard \email{cgid@novonordisk.com}
  \item Kristian, Troejelsgaard ktqn@novonordisk.com
  \item Lovemore Gakava \email{lvgk@novonordisk.com}
  \item Steffen Falgreen Larsen \email{sffl@novonordisk.com}
  \item Vladimir Obucina \email{vlob@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
